Create question 1(lab1)
The values assigned to the data types Char, Short and Int (unsigned and signed) show either an underflow or an overflow from the range of values assigned to the datatypes. When the values surpass the limit, an overflow is reflected. When they go below the limit, an underflow is reflected. 
According to the source code,
The signed Char values have exceeded the -128 to +127 range
The unsigned Char values exceeded the 0 to 255 range

Data type char	variables exceeded the following ranges:
unsigned char	 0 to 255
signed char	-127 to 127
The assigned variables of datatype int exceeded the following ranges:
unsigned int	 0 to 4294967295
signed int	 -2147483648 to 2147483647

The assigned variables of datatype short exceeded the following ranges:
unsigned short int 0 to 65,535
signed short int	 -32768 to 3276